+++ src/main/java/org/nrg/containers/model/command/entity/CommandInputEntity.java
@@ -5,6 +5,7 @@ import org.hibernate.envers.Audited;
 import org.nrg.containers.model.command.auto.Command;
 
 import javax.annotation.Nonnull;
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
@@ -99,6 +100,7 @@ public class CommandInputEntity implements Serializable {
         this.name = name;
     }
 
+    @Column(columnDefinition = "TEXT")
     public String getDescription() {
         return description;
     }
@@ -137,6 +139,7 @@ public class CommandInputEntity implements Serializable {
         this.matcher = matcher;
     }
 
+    @Column(columnDefinition = "TEXT")
     public String getDefaultValue() {
         return defaultValue;
     }