+++ src/main/java/org/nrg/containers/model/command/entity/CommandWrapperExternalInputEntity.java
@@ -5,6 +5,7 @@ import org.hibernate.envers.Audited;
 import org.nrg.containers.model.command.auto.Command;
 
 import javax.annotation.Nonnull;
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -126,6 +127,7 @@ public class CommandWrapperExternalInputEntity {
         this.name = name;
     }
 
+    @Column(columnDefinition = "TEXT")
     public String getDescription() {
         return description;
     }
@@ -174,6 +176,7 @@ public class CommandWrapperExternalInputEntity {
         this.viaSetupCommand = viaSetupCommand;
     }
 
+    @Column(columnDefinition = "TEXT")
     public String getDefaultValue() {
         return defaultValue;
     }